+++ libWebAppServer/proxyTcpServer/proxyServerConn.cc
@@ -178,7 +178,7 @@ proxyServerConn :: onMessage(const WebAppServer::WebAppMessage &m)
                     }
                 }
                 //else cc > 0
-                if (cc != bytes_left)
+                if (cc != (int) bytes_left)
                 {
                     printf("NOTE : write of %u != %d\n",
                            (unsigned int) bytes_left, cc);
